% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\name{addition}
\alias{addition}
\title{addition}
\usage{
addition(value_1, value_2)
}
\arguments{
\item{value_1}{number of the first value}

\item{value_2}{number of the second value}
}
\value{
the sum of the two input numbers
}
\description{
addition
}
\examples{
addition(value_1 = 1, value_2 = 2)
}
