% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hey.R
\name{hey}
\alias{hey}
\title{hey function}
\usage{
hey(name, age, fav_food)
}
\arguments{
\item{name}{A character containing the input name}

\item{age}{A number of age}

\item{fav_food}{A character containing the input of favorite food}
}
\value{

}
\description{
hey function
}
\examples{
hey("Daniel", 31, "pizza")
}
