% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication.R
\name{multiplication}
\alias{multiplication}
\title{Multiplication}
\usage{
multiplication(value_1, value_2)
}
\arguments{
\item{value_1}{first number input}

\item{value_2}{second number input}
}
\value{
product of two input numbers
}
\description{
Multiplication
}
\examples{
multiplication(value_1 = 3, value_2 = 4)
}
